+++ src/utils/rmdirAsync.ts
@@ -1,0 +1,4 @@
+import { promisify } from 'util';
+import { rmdir as rmdirCb } from 'fs';
+
+export const rmdirAsync = promisify(rmdirCb);